+++ src/main/java/com/secuconnect/client/model/PaymentContext.java
@@ -21,6 +21,9 @@ public class PaymentContext {
   @SerializedName("creditcard_schemes")
   private List<String> creditcardSchemes = null;
 
+  @SerializedName("accrual")
+  private Boolean accrual = false;
+
   public PaymentContext autoCapture(Boolean autoCapture) {
     this.autoCapture = autoCapture;
     return this;
@@ -105,6 +108,23 @@ public class PaymentContext {
     this.creditcardSchemes = creditcardSchemes;
   }
 
+  public PaymentContext accrual(Boolean accrual) {
+    this.accrual = accrual;
+    return this;
+  }
+
+   /**
+   * sets accrual flag to payment transaction when created
+   * @return accrual
+  **/
+  public Boolean getAccrual() {
+    return accrual;
+  }
+
+  public void setAccrual(Boolean accrual) {
+    this.accrual = accrual;
+  }
+
   @Override
   public boolean equals(Object o) {
     if (this == o) {
@@ -117,12 +137,13 @@ public class PaymentContext {
     return Objects.equals(this.autoCapture, paymentContext.autoCapture) &&
         Objects.equals(this.paymentMethods, paymentContext.paymentMethods) &&
         Objects.equals(this.merchantInitiated, paymentContext.merchantInitiated) &&
-        Objects.equals(this.creditcardSchemes, paymentContext.creditcardSchemes);
+        Objects.equals(this.creditcardSchemes, paymentContext.creditcardSchemes) &&
+        Objects.equals(this.accrual, paymentContext.accrual);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(autoCapture, paymentMethods, merchantInitiated, creditcardSchemes);
+    return Objects.hash(autoCapture, paymentMethods, merchantInitiated, creditcardSchemes, accrual);
   }
 
   @Override
@@ -133,6 +154,7 @@ public class PaymentContext {
     sb.append("    paymentMethods: ").append(toIndentedString(paymentMethods)).append("\n");
     sb.append("    merchantInitiated: ").append(toIndentedString(merchantInitiated)).append("\n");
     sb.append("    creditcardSchemes: ").append(toIndentedString(creditcardSchemes)).append("\n");
+    sb.append("    accrual: ").append(toIndentedString(accrual)).append("\n");
     sb.append("}");
     return sb.toString();
   }