+++ CHANGELOG.md
@@ -19,6 +19,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 
+## [3.11.0] - 2022-07-14
+[3.10.1]:https://github.com/secuconnect/secuconnect-java-sdk/compare/3.10.1...3.11.0
+
+### Added
+- `Api.GeneralContractsApi.terminate()`
+- `Model.PaymentContext` element `accrual`
+
+
 ## [3.10.1] - 2022-06-21
 [3.10.1]:https://github.com/secuconnect/secuconnect-java-sdk/compare/3.10.0...3.10.1